lsp-devtools: Implement completion item selection
This introduces the `EditorView` widget. It's responsible for managing
the `TextEditor` widget and any additional UI elements e.g. completion
suggestions.

This commit tries to lean on the textual framework more, using
`self.run_worker` for communicating with the LSP server in the
background and a custom message for triggering the display of
completion items
