+++ lib/lsp-devtools/changes/108.enhancement.rst
@@ -1,0 +1,1 @@
+It is now possible to select completion items in the ``lsp-devtools client``
